% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotBLP.R
\name{BLP_plot}
\alias{BLP_plot}
\title{Plot the \eqn{\beta_2} estimates of the Best Linear Predictor}
\usage{
BLP_plot(model, name, rownames = rownames)
}
\arguments{
\item{model}{An `evaluCATE` object}

\item{name}{A string of the plot title}

\item{rownames}{A list of row names to be plotted (from the `evaluCATE` package)}
}
\value{
A plot of the \eqn{\beta_2} estimates of BLP
}
\description{
Plot the \eqn{\beta_2} estimates of the Best Linear Predictor
}
\examples{
BLP_plot(evaluation_tl, "T-Learner RF")
}
