% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DRLearner.R
\name{drlearner}
\alias{drlearner}
\title{DR-Learner}
\usage{
drlearner(X_train, Y_tilde, X_test)
}
\arguments{
\item{X_train}{A vector or a matrix of the covariates in training sample.}

\item{Y_tilde}{A vector of the pseudo-outcomes in training sample.}

\item{X_test}{A vector or matrix of the covariates in test sample.}
}
\value{
A list containing a vector of numeric values of CATE predictions in test sample,
and a fitted weighted regression tree model.
}
\description{
This function predicts the CATE using the DR-Learner.

The DR-Learner uses the pseudo-outcome and applies it in a generic Machine Learning setting.
This specific function applies the DR-Learner in a regression forest setting.

\eqn{\hat{\tau}^{drl}(X) = \underset{\tau}{\operatorname{argmin}} \sum_{i=1}^{N} \bigg(\tilde{Y}_{i, ATE} - \tau(X_i) \bigg)^2}

As a first step, the function runs a regression forest model on the covariates and pseudo-outcomes
in the training sample. Subsequently, it predicts CATEs in test sample.
}
\examples{
drlearner(X_train, Y_tilde, X_test)
}
