% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loop_helper.R
\name{loop_helper}
\alias{loop_helper}
\title{Helper function for Loops}
\usage{
loop_helper(seeds, split)
}
\arguments{
\item{seeds}{A list of seeds to loop over}

\item{split}{A string vector for the train/test split fraction - as title for graph outputs}
}
\value{
A grid of plots
}
\description{
Helper function to run a loop over several seeds to explore the variability of CATE estimates.
}
\examples{
loop_helper(seeds = c(1, 2, 3))
}
