% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RLearner.R
\name{rlearner}
\alias{rlearner}
\title{R-Learner}
\usage{
rlearner(X_train, W_train, Y_train, X_test, mhat, ehat)
}
\arguments{
\item{X_train}{A vector or a matrix of the covariates in training sample.}

\item{W_train}{A vector of the treatment indicator in training sample.}

\item{Y_train}{A vector of the outcomes in training sample.}

\item{X_test}{A vector or matrix of the covariates in test sample.}

\item{mhat}{A vector of estimates of the conditional expectation of the outcome (nuisance parameter)}

\item{ehat}{A vector of estimates of the conditional expectation of the treatment; estimated propensity score (nuisance parameter)}
}
\value{
A vector of numeric values of CATE predictions in test sample.
}
\description{
This function predicts the CATE using the R-Learner.

Every supervised ML model that is capable of dealing with weighted minimization
problems can be applied to the R-Learner. This function uses weighted regression forests and pseudo-outcomes.

The function first calculates the outcome and treatment residuals, then computes the pseudo-outcome by
dividing the outcome residual by the treatment residual. Subsequently, it creates the weights by squaring
the treatment residuals. Finally it fits a weighted regression tree on the covariates in the train sample and the
pseudo outcome and estimates CATE for the test sample.
}
\examples{
rlearner(X_train, W_train, Y_train, X_test, mhat, ehat)
}
