% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TLearner.R
\name{tlearner}
\alias{tlearner}
\title{T-Learner}
\usage{
tlearner(X_train, W_train, Y_train, X_test)
}
\arguments{
\item{X_train}{A vector or a matrix of the covariates in training sample.}

\item{W_train}{A vector of the treatment indicator in training sample.}

\item{Y_train}{A vector of the outcomes in training sample.}

\item{X_test}{A vector or matrix of the covariates in test sample.}
}
\value{
A list of numeric values of CATE predictions in test sample.
}
\description{
This function predicts the CATE using the T-Learner.

It first uses regression forest to fit model \eqn{\hat{m}(1,X)} in treated in train sample.

Then, it uses regression forest to fit model \eqn{\hat{m}(0,X)} in controlled in train sample.

Finally, it estimates CATE as

\eqn{\hat{\tau}(x) = \hat{m}(1,x) - \hat{m}(0,x)}

in test sample
}
\examples{
tlearner(X_train, W_train, Y_train, X_test)
}
